Port `concurrent_streams` test to poll-api
